document_rewriter_flask/requirements.txt
Flask==3.0.3
openai==1.1.1
langchain==0.1.16
langchain-community==0.0.16
python-dotenv==1.0.1
docx2txt==0.8
PyPDF2==3.0.1